pyproject.toml
[project]
name = "notebooklm-mcp"
version = "0.1.0"
description = "MCP server for Google NotebookLM - Research, RAG, and Content Generation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.0.0",
    "notebooklm-py[browser]>=0.3.2",
]

[project.scripts]
notebooklm-mcp = "notebooklm_mcp.server:main"

[build-system]
requires = ["uv_build>=0.10.5,<0.11.0"]
build-backend = "uv_build"
